v3.20.0 Vote for the next map #minor
v3.20.0 Vote for the next map #minor

```diff
+ Features 
```

- **Next map vote system:**
I integrated the vote map system that I've been using for a while on the HLL France servers. 
You'll find several config parameters that allow you to tweak the maps that are suggested by the vote map.
In order to vote a player must write the map number in the chat with: !votemap, e.g: !votemap 1.

You have to add the broadcast showing the map selection your self using your auto broadcast, there are several different formatting style. The ones I've been using are those:

```
120 {scrolling_votemap}
120 Vote for the next map by writing !votemap <map number in the chat>:\n{votenextmap_vertical}
```

Out of the box there's only 1 single reminder that pops up after 20min in game as a welcome message and 5min later restores your original welcome message. This is not really enough on its own. https://prnt.sc/114zai1

When someone vote there's a little feedback, however it may not always show if the autobroad cast happens at the exact same time: https://prnt.sc/114zbfk 
The feedback message stays 10 seconds and then the previous broadcast is restored.

The map selection is automatically generated at the begining of each games and the votes are reset. Everytime someone votes the map rotation is automatically adjusted to put the next map to the one with the most vote. As a safety measure all the maps are included in the rotation in addition of the winning map, so that in case of a bug your server does not end up in a loop.
There are no minimum votes, however thing you somewhat control what is offered there are no surprise.

```diff
+ Features 
```

- Show current welcome message and broadcast (if CRcon was used to set them, it won't work if you use another tool) https://prnt.sc/114z31v
- UX improvement. More space for your test in: Auto broadcast, standard message and puntion reason: https://prnt.sc/114z5ag https://prnt.sc/114z627
- Camera Notifications using broadcast and welcome are now temporary and will restore your original messages after 1 minute
- The map recorder has been fixed to record map restarts properly, your raw map history is available at: http://<url>/api/get_map_history?pretty=1
- Fixed the menu to change Rcons for multi server that use domains so that the path is preserved, so it stays in the same page, e.g you are on the settings page, switch to server 2 will show the settings page.