Squads auto moderation & Perf improvements (#109)
v4.2 Squad Auto Moderation - Perf improvements - Bug fixes #minor

Perf:
- Improved the performance of the Game View by using asynchronous IOs over 20 threads. Note that this significantly slows down the startup of every process (your API, services and such) (as it need to initialise 20 connections to your game server per process). Also reduced the caching time of the role info so that the Game view information shouldn't be older than ~5 seconds. The load time of the Game view is now MUCH faster and therefore more accurate w.r.t. live. The UI now auto-refreshes every 15seconds instead of 30.
- Improved performance of the get_players function (the main one used in your Live view) by using more batch calls. 
- Improved re-ingesting of old logs in case of cache of a cleared / empty cache (redis). It was slowed down by all the errors thrown when trying to get the steamid of players that are not on the server anymore. It's a bit faster now but will still take about 30min / 1h to catch up on the last 180min of logs you get at startup. For that reason it's recommended you live your Rcon live 24/7. 

Bug Fix:
- The scorebot would always recreate a new message everytime its restarted because the path to its DB was not set correctly. 

New feature:  Auto moderation of squads without an Officer (Squad automod for short)

This is a new service that is quite configurable to your taste, please read the config/default_config.yml for more details, I wrote a lot of explainations in there.

In short this is a new "bot" that checks the game view every 5 seconds. It will detect squads that don't have an officier and go through a series of progressive step to resolve the situation until the squad gets an officer and it's probation ends. 
There are up to 3 steps (you chose what you want) that happen one after the other in a escalation manner: 
1. Warn the squads via a broadcast message N times, with a N seconds wait between each warnings
2. Punish the squad members N (or inifinity) times, with a N seconds wait between each warnings (you can make roles and lvl immunte to that)
3. Kick the members

You can chose to skip 1. if you want, however out of fairness for the players you mush at least have one punish configured if you want to have the kick enabled.

Co-authored-by: Julien Marechal <julien.marechal@blah.com>
Co-authored-by: Florian <florian.schmidt.welzow@t-online.de>