Fix historical log download parameter typo (#126)
The player name parameter was incorrectly set to `this.statename` rather than `this.state.name` causing the historical log CSV download to not filter by player name. The data table would show the correct results but on downloading the resulting POST request would return results for all players.

I tested the change on our event server that has access to historical data and it appeared to work correctly.