Add a slider to toggle live log actions between inclusivity and exclusivity.
The live logs view previously would only allow an admin to select specific actions
to view. If an admin wanted everything except one or two things they had to select
a large number of actions every time they used this view.

The slider allows an admin to change between which behavior they prefer, it
defaults to the previous (expected behavior).