Refactor live log action inclusivity filter
Refactor from a toggling padlock control to a select control for increased
clarity and rename associated variables to make it more clear they control
the action filter being either inclusive or exclusive.

Resized the live log controls for the large size breakpoint so they fit
into the 12 column layout and shrunk the refresh control to only be an
icon to save space.

I did not use the Selector component as I could not get it to work correctly
with the text labels and boolean values.

I tested the functionality on our event server and it works.

I tested the UI on different window sizes (on a 1920x1080 17" laptop) and
it's usable and consistent with the previous version of the UI.