Add team balance/shuffle features as API command
The UI for this is not intended for production, it is in a minimal
state for testing purposes only and needs to be refined.

Validation of the API calls is performed in the balance and shuffle
functions which raise errors that are handled in rconweb/api/team_balance.py

This reverts changes from when it was set up as a CLI command.

USERGUIDE.md was updated to document how the features work and
includes warnings about dangerous things to note.

An additional optional webhook is added and the docker compose
files are updated to expose it for multiple servers.

A specific log file (team_balance.log) is added to allow admins to see
any errors or info without having to filter through huge amounts of data
from api.log

I added some tests, but there is not 100% test coverage. Some
tests requires hypothesis to run.

dev-requirements.txt was added to add hypothesis