Fix rcon mod wrongfully considering kill as "during seeding"
In rare cases, the rcon tool might not yet recognized that the server
reached enough population to disable seeding rules (something around 2-7
seconds, depending on how long it takes for the first call to
get_team_view).

Kills during this time are handled by an event, while the seeding
automod assumes the server did not reach the seeding threshold yet. This
might result in kills being punished as they were made with a
seeding-disallowed-weapon, despite the server having enough players.

This commit fixes this by giving automods generally a bit of time after
restaret of the rcon tool to warmup. This might defer or miss
enforcements on some kills, but at least no-one is punished wrongfully.