do not allow same thread to modify idle connections at the same time
For some reason the same thread seems to use more or less concurrently
calls with_connection which might end up in two separate commands in
ServerCtl get the same connection from the pool, resulting in garbage
data for each of the commands. Somehow, this also seems to affect
subsequent requests for a connection to the pool.

It's pretty hard to verify that in a local setup, however, after having
a more explicit locking of the idles connection array with a
conventional (non-reentrant) lock, I did not experience any issues so
far.

Side effect: The lock should also be released when a thread is waiting
for connection to be returned to the pool. At least this will allow
potential other requests to be served, e.g. if a connection errors and a
new one can be opened, instead of "all" waiting for one request to be
served by the pool.