remove explicit reconnect method
It's up to the pool what to do with aconnection that is indicated as
broken by a caller. Right now, simply discarding it is the easiest
solution and done by the pool. A new connection will be established
eventually, if needed. That is basically the same as done by the old
reconnect method.

Reconnecting as it was done right now does not even work, as a closed
socket cannot used to connect to anywhere again.