always release lock, even when connection is not established
If connecting to the HLL server for a new connection fails for whatever
reason, the lock should also be released. Otherwise, the exception will
raise up the call chain, leaving the lock in an unrecoverable state
acquired, which prevents any subsequent request to the pool to
indefinetly wait for the lock to be available.

Also: Never wait indefinetly for a lock, which will never throw any
error in such cases, which makes it especially hard to debug such things
with production logs only.