Always use local time zone in frontend
Instead of using mixed UTC and local times, the frontend should use one timezone consistently.
As local seems to be the more natural thing, I choosed to change all time usages I could find
to the local timezone.
This includes the log view, which might be the most visible change with this commit (the time
in the log view was UTC so far).