Salvage stats when player reconnects
Previously, when a player disconnects and then reconnects to the server
the stats gained in the previous session where lost. The CRCON would
always only save the 'higher' stats value to compensate this a bit, which
was far from perfect.
With this commit, the CRCON will try to salvage the previously gained stats
and combine them after the round ended.