Bugfix: include is_vip and profile in get_detailed_player_info (#715)
`get_detailed_player_info` did not override the default `false` for `is_vip` and `null` for `profile` with the correct values. 
This PR fixes that and removes the workaround in `get_team_view` to not fetch the profiles twice. This also fixes the wrong value `is_vip` in `get_detailed_players`.