replaced emoji-mart with lush-emoji-picker
Due to some unknown reasons the emoji-mart library does not show any other emoji set apart from native one. The problem is that native emoji set does not contain nation flag emojis hence the need for replacement. Lush picker can be much slower and it would be favorable to find the issue with emoji-cart and use that library instead. For the lack of knowledge and to speed up the first release this is the current and hopefuly temporary solution.