Allow customizable CSRF token and session cookie domains
Pass through the CSRF_COOKIE_DOMAIN and SESSION_COOKIE_DOMAIN options as an
environment variable.  This is passed to the backend and read as other options are.

This is particularly useful for RCON's hosted on multiple subdomains sharing login
credentials since the CSRF token or sessionid from one RCON instance can be used on another.

A potential pitfall of this option is if other applications are using CSRF or sessionids with
wildcard domains then they could potentially conflict/cause false errors for logins that are
not the RCON itself. To solve that, we'd have to make our csrftoken and sessionids unique to
the RCON which seems to be a rare and unusual case for most users of the RCON.