Include next_map_id in RCon chat commands to skip current map
This can be used in every place where message variables are supported.
With this commit, these variables are now also available in rcon chat
commands. Most of them are potentially pretty useless, however, the new
next_map_id might be helpful for cases where one wants to easily skip
the current map to the next one in rotation. Either on a public server
or on an event server (where only one map is in the rotation).

A command config should look like this:
{
  "words": [
    "!test2"
  ],
  "description": "Testing layout",
  "commands": {
    "set_map": {
      "map_name": "{next_map_id}"
    }
  },
  "enabled": true,
  "conditions": {}
}