WIP: Use hll-geofence for enforce cap fight
This is a work-in-progress to demonstrate the usage of hll-geofence for
the enforce cap fight seeding automod. It uses the rconv2 protocol and
is therefore much more responsive in detecting players out of bounds.
Usually a player is notified about being out of the play-able area
within 1 second or much less.

This commit introduces a new component that is entirely written in
golang but also requires access to the configuration data in the db.
Therefore there is an added way of an internal API user, which can
authenticate to the CRCon api and use any available api without nay
explicit grants of permissions.

This is a WIP, as it will not take into account saving the seeding
automod config, yet (e.g., by restarting the enforce_cap_fight program
in supervisor). Hence, when changing the config right now, one would
need to restart the program manually.

It does also not yet remove the "offensive-point-tracking" version of
the enforce cap fight component in the python seeding automod.