"none" like rconv1 instead of None for team
Old rconv1 used to have the string "none" as the player team when they were waiting
in the lobby and had not picked a team yet.  Using the keyword None here
instead results in the player being left out of the results from get_team_view
which means the player does not show in the player table in live view and is shown as offline on their player profile.