Fix live view kpm stat (#1138)
Currently kpm on live view uses `current_playtime_seconds` from a
players profile, which for the first match they play in works fine, but
once the map flips and the kills reset this results in the kpm shown
being very low and incorrect.  Instead limit the time used to the lower
of `current_playtime_seconds` or the time since the match started.

Signed-off-by: David Rosendorf <rosendorf.dev@gmail.com>
Co-authored-by: David Rosendorf <rosendorf.dev@gmail.com>